logistic_regression/requirements.txt
pandas
plotly_express
wordcloud
matplotlib.pyplot
pathlib